Add files via upload
Tried to add collision but it really doesn't work. I don't understand the math very well so I can't troubleshoot it well. Also we need a better way to do vector math, probably numpy; things got real slow when the number of balls was high.

Kind of fun to watch though. The big particles tend to go slow and the small ones get going really fast, which I think is the desired behavior.